# API_SPECIFICATION.yaml
openapi: 3.0.0
info:
  title: Dhanam Ledger API
  version: 1.0.0
  description: Comprehensive financial management API for personal and business budgeting
  contact:
    name: Dhanam Support
    email: support@dhanam.app
  license:
    name: MIT
servers:
  - url: https://api.dhanam.app/v1
    description: Production server
  - url: https://staging-api.dhanam.app/v1
    description: Staging server
  - url: http://localhost:4000/v1
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /spaces:
    get:
      tags: [Spaces]
      summary: List user spaces
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        200:
          description: List of spaces
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedSpaces'

    post:
      tags: [Spaces]
      summary: Create new space
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, type]
              properties:
                name:
                  type: string
                type:
                  type: string
                  enum: [personal, business]
                currency:
                  type: string
                  enum: [MXN, USD, EUR]
                  default: MXN
      responses:
        201:
          description: Space created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Space'
  /spaces/{spaceId}:
    get:
      tags: [Spaces]
      summary: Get space details
      parameters:
        - $ref: '#/components/parameters/SpaceIdParam'
      responses:
        200:
          description: Space details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Space'
        404:
          $ref: '#/components/responses/NotFound'

  # Accounts
...
  /spaces/{spaceId}/transactions:
    get:
      tags: [Transactions]
      summary: List transactions
      parameters:
        - $ref: '#/components/parameters/SpaceIdParam'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: accountId
          in: query
          schema:
            type: string
        - name: categoryId
          in: query
          schema:
            type: string
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
        - name: search
          in: query
          schema:
            type: string
      responses:
        200:
          description: List of transactions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedTransactions'

    post:
      tags: [Transactions]
      summary: Create manual transaction
      parameters:
        - $ref: '#/components/parameters/SpaceIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [accountId, amount, date, description]
              properties:
                accountId:
                  type: string
                amount:
                  type: number
                currency:
                  type: string
                description:
                  type: string
                categoryId:
                  type: string
                date:
                  type: string
                  format: date
                pending:
                  type: boolean
                  default: false
      responses:
        201:
          description: Transaction created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
...
  /spaces/{spaceId}/budgets:
    get:
      tags: [Budgets]
      summary: List budgets
      parameters:
        - $ref: '#/components/parameters/SpaceIdParam'
        - name: active
          in: query
          schema:
            type: boolean
      responses:
        200:
          description: List of budgets
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Budget'

    post:
      tags: [Budgets]
      summary: Create budget
      parameters:
        - $ref: '#/components/parameters/SpaceIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, period, startDate, categories]
              properties:
                name:
                  type: string
                period:
                  type: string
                  enum: [monthly, quarterly, yearly]
                startDate:
                  type: string
                  format: date
                endDate:
                  type: string
                  format: date
                categories:
                  type: array
                  items:
                    type: object
                    required: [categoryId, amount]
                    properties:
                      categoryId:
                        type: string
                      amount:
                        type: number
                        minimum: 0
                      alertThreshold:
                        type: number
                        minimum: 0
                        maximum: 100
                        default: 80
      responses:
        201:
          description: Budget created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Budget'

  # Analytics
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    SpaceIdParam:
      name: spaceId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  schemas:
...
    Space:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        type:
          type: string
          enum: [personal, business]
        currency:
          type: string
        role:
          type: string
          enum: [owner, admin, member, viewer]
        createdAt:
          type: string
          format: date-time
...
    Transaction:
      type: object
      properties:
        id:
          type: string
          format: uuid
        accountId:
          type: string
          format: uuid
        amount:
          type: number
        currency:
          type: string
        description:
          type: string
        category:
          $ref: '#/components/schemas/Category'
        date:
          type: string
          format: date
        pending:
          type: boolean
        metadata:
          type: object
        createdAt:
          type: string
          format: date-time
    Category:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        icon:
          type: string
        color:
          type: string
        parentId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
...
    Budget:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        period:
          type: string
          enum: [monthly, quarterly, yearly]
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        categories:
          type: array
          items:
            type: object
            properties:
              category:
                $ref: '#/components/schemas/Category'
              budgeted:
                type: number
              spent:
                type: number
              remaining:
                type: number
              percentUsed:
                type: number
              alertThreshold:
                type: number
        totalBudgeted:
          type: number
        totalSpent:
          type: number
        totalRemaining:
          type: number
        createdAt:
          type: string
          format: date-time
...
    PaginatedSpaces:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Space'
        pagination:
          $ref: '#/components/schemas/Pagination'
    PaginatedTransactions:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Transaction'
        pagination:
          $ref: '#/components/schemas/Pagination'
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        pages:
          type: integer
        hasNext:
          type: boolean
        hasPrev:
          type: boolean
    ApiError:
      type: object
      properties:
        success:
          type: boolean
          default: false
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: array
              items:
                type: object
        meta:
          type: object
          properties:
            timestamp:
              type: string
              format: date-time
            requestId:
              type: string
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...